Fix: modificar tipo do id postagem

+++ src/main/java/br/com/zup/gerenciadorDePostagem/postagem/PostagemController.java
@@ -45,7 +45,7 @@ public class PostagemController {
     }
 
     @PutMapping("/{id}")
-    public void editarPostagem(@PathVariable int id, @RequestBody @Valid PostagemDTO postagemDTO,
+    public void editarPostagem(@PathVariable Long id, @RequestBody @Valid PostagemDTO postagemDTO,
                                Authentication authentication) {
 
         UsuarioLogado usuarioLogado = (UsuarioLogado) authentication.getPrincipal();
